Add app.rb file and people_creation.rb module to comply with single responsability principle
